HawkAI/agents/analytics_agent.py
from google.adk.agents import Agent
from typing import Dict, Any


# agents/analytics_agent.py
class DataAnalyticsAgent(Agent):
    """Agent specialized in data analysis and pattern recognition"""
    
    def __init__(self, project_id: str, location: str):
        super().__init__(
            name="DataAnalytics-Agent", 
            description="Specialized agent for data analysis, pattern recognition, and predictive modeling",
            instructions="You analyze historical data, detect patterns and anomalies, and provide predictive insights."
        )
        vertexai.init(project=project_id, location=location)
        
        self.model = GenerativeModel(
            model_name=MODEL_NAME_FLASH_2,
            system_instruction="""
            You are a Data Analytics Agent specialized in:
            - Historical incident pattern analysis
            - Predictive modeling for event safety
            - Statistical trend identification
            - Anomaly detection in event data
            - Performance metrics calculation
            
            Always provide data-driven insights with confidence levels.
            """
        )
        self.chat_session = self.model.start_chat()
    
    def analyze_historical_patterns(self, incident_data: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze historical incident patterns"""
        prompt = f"""
        Analyze these historical incident patterns:
        Data: {incident_data}
        
        Identify:
        1. Common incident types and frequencies
        2. Time-based patterns (day/hour correlations)
        3. Location hotspots
        4. Predictive indicators
        
        Provide structured analysis with confidence scores.
        """
        
        response = self.chat_session.send_message(prompt)
        return {
            "agent": "data_analytics",
            "analysis_type": "historical_patterns",
            "result": response.text,
            "data_points": len(incident_data.get('incidents', []))
        }
    
    def detect_anomalies(self, current_metrics: Dict[str, Any]) -> Dict[str, Any]:
        """Detect anomalies in current event metrics"""
        prompt = f"""
        Detect anomalies in current event metrics:
        Current metrics: {current_metrics}
        
        Compare against typical patterns and identify:
        - Unusual crowd behavior patterns
        - Abnormal resource utilization
        - Unexpected metric correlations
        
        Rate anomaly severity and provide investigation priorities.
        """
        
        response = self.chat_session.send_message(prompt)
        return {
            "agent": "data_analytics",
            "analysis_type": "anomaly_detection",
            "result": response.text,
            "metrics_analyzed": list(current_metrics.keys())
        }