HawkAI/requirements.txt
functions-framework==3.8.1
vertexai==1.60.0
google-cloud-logging==3.8.0
Pillow==10.1.0
